fix: double-quote string mode
needed to match a single double-quote character (")"),
without three of them (""") also being matched.

Originally tried to solve this with fancy regexes that matched a " and
two non-" chars, but this kept breaking
(they kept consuming each other's non-" chars)

So instead, just used rule positioning so a """ always matches first.
So if it is a """, it will be dealt with first.