fix: single-word replacements that had a \b before it
this doesn't work in Transcribers, as the string the rule operates on
is always substring()ed so the next unconsumed char is at index 0.

The start of a string always matches \b,
so this is why "\bor\b" was matching "for". The f was already cut off.

TODO: turn this into its own rule.
It's always the same \b simulation regex for the consumed matcher,
so we could just make that a part of another IRule subclass.