feat: first attempt at indents to braces
this is a flawed algorithm, because python can end multiple indents in one go,
requiring possibly multiple closing braces.

We need a stack-based approach to close as many indents as we have seen.

But this is good work implementing the algorithm as designed,
and is a good starting point for the stack-based version